fix: Fixing unexpected cutout by only setting person detected to false once floor is detected while person_detected is set to True

+++ api/app.py
@@ -3,7 +3,7 @@ from datetime import datetime as dt
 import os
 import re
 from datetime import datetime as dt
-from flask import Flask,  request, jsonify
+from flask import Flask, appcontext_tearing_down,  request, jsonify
 from flask_restful import Api
 from flask_cors import CORS
 from threading import Thread
@@ -71,45 +71,46 @@ class ExhibitArea:
         # self.motion_id = id
         self.person_detected = False
         self.proc = None
-        self.floor_ranges = []
+        self._floor_ranges = []
+        self.floor_range = -1
         self.calibrate_range(approx_floor_height)
         self.last_range = approx_floor_height
 
     def calibrate_range(self, floor_range_reading):
-        self.floor_ranges.append(floor_range_reading)
-        self.floor_ranges = sorted(self.floor_ranges)
-        self.trigger_range = self.floor_ranges[len(
-            self.floor_ranges) // 2] - 20
+        self._floor_ranges.append(floor_range_reading)
+        self._floor_ranges = sorted(self._floor_ranges)
+        self.floor_range = self._floor_ranges[len(
+            self._floor_ranges) // 2]
+        self.trigger_range = self.floor_range - 20
 
     def set_person_detected(self, person_detected):
         self.person_detected = person_detected
-        if self.person_detected == False:
-            self.fade_out()
+        if self.proc and self.person_detected == False:
+            self.proc.terminate()
 
     def play_audio(self):
         global playlist
 
         artist = get_artist_by_time(dt.now())
+        # artist = "4"
         channel = int(self.sound_id) - 1
         audio_file = playlist[artist][channel]
 
-        cmd = f'ffmpeg -i {audio_file} -ac 16 -filter_complex "[0:a]pan=16c|c{channel}=c0[a];[a]dynaudnorm=p=0.9:s=5[a_norm]" -map "[a_norm]" -f audiotoolbox -audio_device_index 1 -'
+        cmd = ""
+        if artist != "4":
+            cmd = f'ffmpeg -i {audio_file} -ac 16 -filter_complex "[0:a]loudnorm=I=-14:LRA=5:TP=-1.5[a];[a]pan=16c|c{channel}=c0[b];[b]volume=1.4[c]" -map "[c]" -f audiotoolbox -audio_device_index 1 -'
+        else:
+            cmd = f'ffmpeg -i {audio_file} -ac 16 -filter_complex "[0:a]loudnorm=I=-14:LRA=5:TP=-1.5[a];[a]pan=16c|c{channel}=c0[b];[b]volume=1.2[c]" -map "[c]" -f audiotoolbox -audio_device_index 1 -'
 
         while self.person_detected:
             audio_log.write(f"{dt.now()}\n{cmd}\n" + ("-"*80) + '\n')
 
-            self.proc = Popen(
-                cmd, shell=True, stdout=FFMPEG_OUT, stderr=FFMPEG_ERR)
-            self.proc.wait()
-
-    def fade_out(self):
-        if self.proc:
-            # # Kill ffmpeg process and fade audio out
-            # cmd = f'ffmpeg -f lavfi -i anullsrc=channel_layout=stereo:sample_rate=44100 -t 1 -filter_complex "[0:a]afade=t=out:st=0:d=1[a]" -map "[a]" -f audiotoolbox -audio_device_index 1 -'
-            # self.fade = Popen(
-            #     cmd, shell=True, stdout=FFMPEG_OUT, stderr=FFMPEG_ERR)
-            self.proc.send_signal(2)
-            self.proc.terminate()
+            if DEVELOPMENT:
+                time.sleep(5)
+            else:
+                self.proc = Popen(
+                    cmd, shell=True, stdout=FFMPEG_OUT, stderr=FFMPEG_ERR)
+                self.proc.wait()
 
 
 area_map = {f"{sound_id}": ExhibitArea(sound_id) for sound_id in sound_sensors}
@@ -117,7 +118,14 @@ area_thread_map = {f"{sound_id}": Thread(
     target=area_map[sound_id].play_audio) for sound_id in sound_sensors}
 
 
+# t = area_thread_map["1"]
+# area_map["1"].person_detected = True
+# t.start()
+# if t.is_alive():
+#     t.join()
+
 # -------------------------------- FLASK API --------------------------------
+
 
 app = Flask(__name__)
 api = Api(app)
@@ -158,20 +166,20 @@ def data():
         thread = area_thread_map[sensorID_]
 
         if CALIBRATION_DONE:
-            if range_ > 0 and range_ <= area.trigger_range:  # received valid range reading
-                if not thread.is_alive():
+            if range_ > 0 and range_ <= area.trigger_range:
+                if not area.person_detected and not thread.is_alive():
                     area.set_person_detected(True)
                     thread.start()
                 # 5cm consistency check first -> abs(range_ - area.last_range) < 5
-                area.last_range = range_
             # elif motion_ == 1:  # received valid motion reading
             #     if not area.person_detected and not thread.is_alive():
             #         area.set_person_detected(True)
             #         thread.start()
-            else:
-                area.set_person_detected(False)
-                area_thread_map[sensorID_] = Thread(
-                    target=area_map[sensorID_].play_audio)
+            elif abs(range_ - area.floor_range) < 5:
+                if area.person_detected:
+                    area.set_person_detected(False)
+                    area_thread_map[sensorID_] = Thread(
+                        target=area_map[sensorID_].play_audio)
         else:
             area.calibrate_range(range_)
             if (dt.now() - start_time).seconds > INIT_CALIBRATION_SECONDS:
